% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS2R.R
\name{SurveySolutions2R}
\alias{SurveySolutions2R}
\title{Read in Survey Data Exported In Tabular Format from Survey Solutions HQ}
\usage{
SurveySolutions2R(zipfile, encoding = "UTF-8", dupLabels = TRUE,
  saveData = FALSE, ...)
}
\arguments{
\item{zipfile}{The path of a zip file produced in the Data Export Menu for the
Main Survey Data in "Tabular Format".}

\item{encoding}{default is "UTF-8". Other possible options are "unknown" and 
"Latin-1". Note: it is not used to re-encode the input, rather enables 
handling of encoded strings in their native encoding.}

\item{dupLabels}{default is TRUE. If TRUE, and the same label is used for 
multiple codes within a factor, then the output file will not distinguish 
between the same code. If FALSE then new unique labels will be created by 
concatenating the label and the code.}

\item{saveData}{default is FALSE. If TRUE then filename will be derived from 
input file location. Can also be an explcit filename.}

\item{...}{Additional import options for \code{\link[data.table]{fread}}}
}
\value{
A series of data frames
}
\description{
Takes a zip file produced by the Data Export Menu from Survey Solutions HQ and
imports all data frames in the R environment, assigns factor levels and adds
full text for variable labels using \code{\link[Hmisc]{label}}
}
\examples{
setwd("C:/Users/sdumb/Dropbox (SSD)/SurveySolutions2R")
SurveySolutions2R("Example-FP_31_Tabular_All.zip",saveData=FALSE)
}
\keyword{Import}
\keyword{Survey}
\keyword{SurveySolutions}
\keyword{Tabular}
